\encoding{UTF-8}
\name{genSurv-package}
\alias{genSurv-package}
\docType{package}
\title{Generating Multi-State Survival Data}
\description{
  The \pkg{genSurv} software permits to generate data wih one binary time-dependent covariate and data stemming from a progressive illness-death model.
}

\details{
	\tabular{ll}{
		Package: \tab genSurv\cr
		Type: \tab Package\cr
		Version: \tab 1.0.3\cr
		Date: \tab 2015-11-09\cr
		License: \tab GPL (>= 2)\cr
		LazyLoad: \tab yes\cr
		LazyData: \tab yes\cr
	}
}

\author{
	Artur Araújo, Luís Meira-Machado \email{lmachado@math.uminho.pt}\cr
  and Susana Faria \email{sfaria@math.uminho.pt}\cr
	Maintainer: Artur Araújo \email{artur.stat@gmail.com}
}

\references{
  Anderson, P. K., Gill, R. D. (1982) Cox's regression model for counting processes: a large sample study. Annals of Statistics, 10:1100-1120.

  Cox, D.R. (1972). Regression models and life tables (with discussion). Journal of the Royal Statistical Society, Series B 34:187-220.

	Jackson, C. (2007). Multi-state modelling with R: the msm package, Cambridge, U.K.

  Johnson, M. E. (1987). Multivariate Statistical Simulation, John Wiley and Sons.

	Johnson, N., Kotz, S. (1972). Distribution in statistics: continuous multivariate distributions, John Wiley and Sons.

  Lu J., Bhattacharya, G. (1990). Some new constructions of bivariate weibull models, Annals of Institute of Statistical Mathematics, 42:543-559.

	Meira-Machado, L., Cadarso-Suárez, C., De Uña- Álvarez, J., Andersen, P.K. (2009). Multi-state models for the analysis of time to event data. Statistical Methods in Medical Research 18(2):195-222.

	Meira-Machado, L., Roca-Pardiñas, J. (2011). p3state.msm: Analyzing Survival Data from an Illness-Death Model. Journal of Statistical Software, 38(3): 1-18.

	Therneau, T.M., Grambsch, P.M. (2000). Modelling survival data: Extending the Cox Model. New York: Springer.
}

\keyword{package}
