\encoding{UTF-8}
\name{genSurv-package}
\alias{genSurv-package}
\docType{package}
\title{Generating Multi-State Survival Data}
\description{
The \pkg{genSurv} software permits to generate data with one binary time-dependent covariate and data stemming from a progressive illness-death model.
}

\details{
	\tabular{ll}{
		Package: \tab genSurv\cr
		Type: \tab Package\cr
		Version: \tab 1.0.4\cr
		Date: \tab 2021-10-19\cr
		License: \tab GPL-3\cr
		LazyLoad: \tab yes\cr
	}
}

\author{
Artur Araújo, Luís Meira-Machado \email{lmachado@math.uminho.pt}\cr
and Susana Faria \email{sfaria@math.uminho.pt}\cr
Maintainer: Artur Araújo \email{artur.stat@gmail.com}
}

\references{
Anderson, P.K., Gill, R.D. (1982). Cox's regression model for counting processes: a large sample study. \emph{Annals of Statistics}, \bold{10}(4), 1100-1120. \doi{10.1214/aos/1176345976}

Cox, D.R. (1972). Regression models and life tables. \emph{Journal of the Royal Statistical Society: Series B}, \bold{34}(2), 187-202. \doi{10.1111/j.2517-6161.1972.tb00899.x}

Jackson, C. (2011). Multi-State Models for Panel Data: The msm Package for R. \emph{Journal of Statistical Software}, \bold{38}(8), 1–28. \doi{10.18637/jss.v038.i08}

Johnson, M. E. (1987). \emph{Multivariate Statistical Simulation}, John Wiley and Sons.

Johnson, N., Kotz, S. (1972). \emph{Distribution in statistics: continuous multivariate distributions}, John Wiley and Sons.

Lu J., Bhattacharya G. (1990). Some new constructions of bivariate weibull models. \emph{Annals of Institute of Statistical Mathematics}, \bold{42}(3), 543-559. \doi{10.1007/BF00049307}

Meira-Machado, L., Cadarso-Suárez, C., De Uña- Álvarez, J., Andersen, P.K. (2009). Multi-state models for the analysis of time to event data. \emph{Statistical Methods in Medical Research}, \bold{18}(2), 195-222. \doi{10.1177/0962280208092301}

Meira-Machado L., Faria S. (2014). A simulation study comparing modeling approaches in an illness-death multi-state model. \emph{Communications in Statistics - Simulation and Computation}, \bold{43}(5), 929-946. \doi{10.1080/03610918.2012.718841}

Meira-Machado, L., Roca-Pardiñas, J. (2011). p3state.msm: Analyzing Survival Data from an Illness-Death Model. \emph{Journal of Statistical Software}, \bold{38}(3), 1-18. \doi{10.18637/jss.v038.i03}

Meira-Machado, L., Sestelo M. (2019). Estimation in the progressive illness-death model: a nonexhaustive
review. \emph{Biometrical Journal}, \bold{61}(2), 245–263. \doi{10.1002/bimj.201700200}

Therneau, T.M., Grambsch, P.M. (2000). \emph{Modelling survival data: Extending the Cox Model}, New York: Springer.
}

\keyword{package}
